:truck: moved `YoutubePlayerScaffold` to separate file
